+++ app/public/js/site.js
@@ -1,0 +1,8 @@
+'use strict';
+
+$(document).ready(function () {
+  $('#fullpage').fullpage({
+    sectionsColor: ['#F3F3F3', '#FFFFFF', '#F3F3F3', '#FFFFFF', '#F3F3F3', '#FFFFFF', '#F3F3F3'],
+    anchors: ['main', 'our-story', 'the-big-day', 'accomodation', 'explore', 'song-requests', 'registry']
+  });
+});